+++ equi/src/main/java/com/example/equi/service/CalcFOT.java
@@ -1,9 +1,11 @@
 package com.example.equi.service;
 
+import java.math.BigInteger;
+
 public class CalcFOT {
-    public static Integer calcSum(Integer hours, Integer rate) {
-        Integer sum;
-        sum = hours * rate;
+    public static BigInteger calcSum(Integer hours, Integer rate) {
+        BigInteger sum;
+        sum = BigInteger.valueOf(hours * rate);
         return sum;
     }
 }
